requirements.txt
fastapi==0.115.6
uvicorn==0.32.1
pydantic==2.9.2
httpx==0.27.2
python-dotenv
